--- a/src/server/api/url_parser.cpp
+++ b/src/server/api/url_parser.cpp
@@ -54,7 +54,8 @@ boost::optional<ParsedURL> parseURL(std::string::iterator &iter, const std::stri
         {
             iter = iter_copy;
             // TODO: find a way to do it more effective
-            std::string parsed_part = "/" + out.service + "/v" + std::to_string(out.version) + "/" + out.profile + "/";
+            std::string parsed_part =
+                "/" + out.service + "/v" + std::to_string(out.version) + "/" + out.profile + "/";
             out.prefix_length = parsed_part.length();
             return boost::make_optional(out);
         }