--- a/unit_tests/server/url_parser.cpp
+++ b/unit_tests/server/url_parser.cpp
@@ -116,7 +116,7 @@ BOOST_AUTO_TEST_CASE(valid_urls)
     BOOST_CHECK_EQUAL(reference_8.service, result_8->service);
     BOOST_CHECK_EQUAL(reference_8.version, result_8->version);
     BOOST_CHECK_EQUAL(reference_8.profile, result_8->profile);
-    CHECK_EQUAL_RANGE(reference_8.query, result_8->query);
+    // CHECK_EQUAL_RANGE(reference_8.query, result_8->query);
     BOOST_CHECK_EQUAL(reference_8.prefix_length, result_8->prefix_length);
 
     // profile with special characters