# Issue

What issue is this PR targeting? If there is no issue that addresses the problem, please open a corresponding issue and link it here.

Please read our [documentation](https://github.com/Project-OSRM/osrm-backend/blob/master/docs/releasing.md) on release and version management.
If your PR is still work in progress please attach the relevant label.

## Tasklist

 - [ ] CHANGELOG.md entry ([How to write a changelog entry](http://keepachangelog.com/en/1.0.0/#how))
 - [ ] update relevant [Wiki pages](https://github.com/Project-OSRM/osrm-backend/wiki)
 - [ ] add tests (see [testing documentation](https://github.com/Project-OSRM/osrm-backend/blob/master/docs/testing.md))
 - [ ] review
 - [ ] adjust for comments
 - [ ] cherry pick to release branch

## Requirements / Relations

 Link any requirements here. Other pull requests this PR is based on?


<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1131.58<br/>plain u32: 1141.42<br/>aliased double: 1170.09<br/>plain double: 1167.46 | aliased u32: 1145.28<br/>plain u32: 1141.12<br/>aliased double: 1178.48<br/>plain double: 1174.34 |
| json-render | String: 8.4204ms<br/>Stringstream: 11.4856ms<br/>Vector: 7.54544ms | String: 8.40765ms<br/>Stringstream: 11.6779ms<br/>Vector: 7.61471ms |
| match_ch | Default radius: <br/>4.46775ms/req at 82 coordinate<br/>0.0544847ms/coordinate<br/>Radius 5m: <br/>4.46646ms/req at 82 coordinate<br/>0.054469ms/coordinate<br/>Radius 10m: <br/>15.1237ms/req at 82 coordinate<br/>0.184436ms/coordinate<br/>Radius 15m: <br/>36.7593ms/req at 82 coordinate<br/>0.448285ms/coordinate<br/>Radius 30m: <br/>313.514ms/req at 82 coordinate<br/>3.82334ms/coordinate | Default radius: <br/>4.47231ms/req at 82 coordinate<br/>0.0545403ms/coordinate<br/>Radius 5m: <br/>4.47536ms/req at 82 coordinate<br/>0.0545776ms/coordinate<br/>Radius 10m: <br/>15.1969ms/req at 82 coordinate<br/>0.185328ms/coordinate<br/>Radius 15m: <br/>36.853ms/req at 82 coordinate<br/>0.449427ms/coordinate<br/>Radius 30m: <br/>314.338ms/req at 82 coordinate<br/>3.83339ms/coordinate |
| match_mld | Default radius: <br/>3.36122ms/req at 82 coordinate<br/>0.0409905ms/coordinate<br/>Radius 5m: <br/>3.4193ms/req at 82 coordinate<br/>0.0416988ms/coordinate<br/>Radius 10m: <br/>12.38ms/req at 82 coordinate<br/>0.150976ms/coordinate<br/>Radius 15m: <br/>31.88ms/req at 82 coordinate<br/>0.388781ms/coordinate<br/>Radius 30m: <br/>367.753ms/req at 82 coordinate<br/>4.48479ms/coordinate | Default radius: <br/>3.37803ms/req at 82 coordinate<br/>0.0411955ms/coordinate<br/>Radius 5m: <br/>3.36832ms/req at 82 coordinate<br/>0.041077ms/coordinate<br/>Radius 10m: <br/>12.3556ms/req at 82 coordinate<br/>0.150677ms/coordinate<br/>Radius 15m: <br/>31.7308ms/req at 82 coordinate<br/>0.386961ms/coordinate<br/>Radius 30m: <br/>351.2ms/req at 82 coordinate<br/>4.28293ms/coordinate |
| packedvector | random write:<br/>std::vector 11190.4 ms<br/>util::packed_vector 73894.5 ms<br/>slowdown: 6.6034<br/>random read:<br/>std::vector 11022.7 ms<br/>util::packed_vector 30269.3 ms<br/>slowdown: 2.74609 | random write:<br/>std::vector 11259.8 ms<br/>util::packed_vector 82037.7 ms<br/>slowdown: 7.28589<br/>random read:<br/>std::vector 11064.5 ms<br/>util::packed_vector 33326.6 ms<br/>slowdown: 3.01204 |
| rtree | 1 result:<br/>207.122ms ->  0.0207122 ms/query<br/>10 results:<br/>242.976ms ->  0.0242976 ms/query | 1 result:<br/>206.74ms ->  0.020674 ms/query<br/>10 results:<br/>241.717ms ->  0.0241717 ms/query |
<!-- BENCHMARK_RESULTS_END -->